requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymysql==1.1.0
SQLAlchemy==2.0.23
geohash2==1.1
pytz==2023.3
werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==12.1.1
requests>=2.28.0
cryptography>=41.0.0
